requirements.txt
Faker==26.0.0
kafka-python==2.0.2
neo4j==5.22.0

